Fixed OSS makefile to install required headers and libs under $(SSTACK_ROOT)/oss_lib
To include OSS headers, use $(SSTACK_ROOT)/oss_libs/include
To link with OSS library, use $(SSTACK_ROOT)/oss_libs/lib

Compilation is through with this change.

One issue that remains is OSS is getting rebuilt each time irrespective of whether
it is built earlier or not. One way to avoid it is to create a file under oss
and not to build if that file exists. This is done through depend.
Will fix it later. For now, things are working.